tests/wpt/web-platform-tests/web-animations/animation/pause.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation.pause()</title>
<link rel="help" href="https://w3c.github.io/web-animations/#dom-animation-pause">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
"use strict";

promise_test(function(t) {
  var div = createDiv(t);
  var animation = div.animate({}, 1000 * MS_PER_SEC);
  var previousCurrentTime = animation.currentTime;

  return animation.ready.then(waitForAnimationFrames(1)).then(function() {
    assert_true(animation.currentTime >= previousCurrentTime,
                'currentTime is initially increasing');
    animation.pause();
    return animation.ready;
  }).then(function() {
    previousCurrentTime = animation.currentTime;
    return waitForAnimationFrames(1);
  }).then(function() {
    assert_equals(animation.currentTime, previousCurrentTime,
                  'currentTime does not increase after calling pause()');
  });
}, 'pause() a running animation');

promise_test(function(t) {
  var div = createDiv(t);
  var animation = div.animate({}, 1000 * MS_PER_SEC);

  // Go to idle state then pause
  animation.cancel();
  animation.pause();

  assert_equals(animation.currentTime, 0, 'currentTime is set to 0');
  assert_equals(animation.startTime, null, 'startTime is not set');
  assert_equals(animation.playState, 'pending', 'initially pause-pending');

  // Check it still resolves as expected
  return animation.ready.then(function() {
    assert_equals(animation.playState, 'paused',
                  'resolves to paused state asynchronously');
    assert_equals(animation.currentTime, 0,
                  'keeps the initially set currentTime');
  });
}, 'pause() from idle');

promise_test(function(t) {
  var div = createDiv(t);
  var animation = div.animate({}, 1000 * MS_PER_SEC);
  animation.cancel();
  animation.playbackRate = -1;
  animation.pause();

  assert_equals(animation.currentTime, 1000 * MS_PER_SEC,
                'currentTime is set to the effect end');

  return animation.ready.then(function() {
    assert_equals(animation.currentTime, 1000 * MS_PER_SEC,
                  'keeps the initially set currentTime');
  });
}, 'pause() from idle with a negative playbackRate');

test(function(t) {
  var div = createDiv(t);
  var animation = div.animate({}, {duration: 1000 * MS_PER_SEC,
                                   iterations: Infinity});
  animation.cancel();
  animation.playbackRate = -1;

  assert_throws('InvalidStateError',
                function () { animation.pause(); },
                'Expect InvalidStateError exception on calling pause() ' +
                'from idle with a negative playbackRate and ' +
                'infinite-duration animation');
}, 'pause() from idle with a negative playbackRate and endless effect');

promise_test(function(t) {
  var div = createDiv(t);
  var animation = div.animate({}, 1000 * MS_PER_SEC);
  return animation.ready
    .then(function(animation) {
      animation.finish();
      animation.pause();
      return animation.ready;
    }).then(function(animation) {
      assert_equals(animation.currentTime, 1000 * MS_PER_SEC,
                    'currentTime after pausing finished animation');
    });
}, 'pause() on a finished animation');

</script>
</body>
